feat(stats): enhance stats support and front
- Added new intervals: HOUR and DAY to the Interval enum.
- Updated RepositoryUtils to throw exceptions for null values in UUID and int conversions.
- Enhanced StatsQueryValidator to validate bucket counts for HOUR and DAY intervals.
- Modified StatsServiceImpl to dynamically generate supported intervals in the catalog response.
- Introduced integration tests for catalog and distribution endpoints to ensure correct functionality.

+++ backend/src/main/java/com/scorpanion/backend/stats/service/internal/ScoreBucketing.java
@@ -16,7 +16,7 @@ public final class ScoreBucketing {
 			return List.of(new ScoreBucket(min, min + 1, String.valueOf(min)));
 		}
 
-		int range = max - min;
+		long range = (long) max - min;
 		double rawStep = (double) range / TARGET_BUCKET_COUNT;
 		int step = niceStep(rawStep);
